pyproject.toml
[project]
name = "torchLoom"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
version = "0.1.0"
dependencies = [
    "torch>=2.7",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "websockets>=12.0",
    "nats-py>=2.6.0",
    "lightning>=2.0.0",
    "grpcio-tools>=1.60.0",
    "protobuf>=4.25.0",
]

[project.urls]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-timeout",
    "black",
    "pyre-check",
    "parameterized",
    "expecttest",
    "pytest-asyncio",
]

[tool.isort]
multi_line_output = 3
combine_as_imports = true

[tool.pytest.ini_options]
log_format = "%(asctime)s %(levelname)s %(message)s"
log_date_format = "%Y-%m-%d %H:%M:%S"
log_level = "INFO"
timeout = 60
timeout_method = "thread"
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["torchLoom"]
